List of all functions of DiaCARS.jl

DiaCARS.calculateLineWidths — Method

calculateLineWidths(T,FJ,species)

Calculates the Linewidths for a given species using the MEG law

using DiaCARS
using Plots

Species = "N2"
T=2000
MolParams = DiaCARS.getMolecularParameters(Species)
TT,FJ,GJ,G0V = DiaCARS.calculateTermEnergies(MolParams)
transitions = DiaCARS.calculateLinePositions(TT)
γN2 = DiaCARS.calculateLineWidths(T,FJ,Species)

Species = "O2"
T=2000
MolParams = DiaCARS.getMolecularParameters(Species)
TT,FJ,GJ,G0V = DiaCARS.calculateTermEnergies(MolParams)
transitions = DiaCARS.calculateLinePositions(TT)
γO2 = DiaCARS.calculateLineWidths(T,FJ,Species)

# Plot the result
plotlyjs()
p=plot(DiaCARS.J,[γN2 γO2],label=["N2" "O2"])
xlabel!("J")
ylabel!("γ FWHM in cm-1")
savefig(p, "figure2.html") # ignore this line, this is just for the documentation build process

<iframe src="../figure2.html" style="height:500px;width:100%;"></iframe>

DiaCARS.dconv_edge — Method

dconv_edge() Function to do a fast convolution on an array signal with kernel kern that produces an output that has the same size as signal. Attention: has edge effects by design, account for this by enlargening your input signal. Uses LoopVectorization for speed.

DiaCARS.simulateConvolvedSpectrum — Method

simulateConvolvedSpectrum(χR,χI,γ,method,ω)

Takes the result of simulateTheoreticalSusceptibility() and performs the convolution. method can be "Yuratich" or "Teets", see Ref XX for what this means. Depending on the size of χ, this function detects if the functions for dual- or single-pump spectra is used.

DiaCARS.simulateTheoreticalSusceptibility — Method

simulateTheoreticalSusceptibility(T,Species,P,Model,ROI,FrequencyOffset)

Simulates the theoretical suscepbtility of a given Species based on temperature T in K, pressure P in bar on a wavenumberarray ω based on the region of interest ROI The calculated amplitudes are valid only for single-pump CARS. For Dual-Pump Spectra, the results have to be divided by two.